allow multiple simultaneous calls to LookupMetadata
LookupMetadata both reads from and writes to a raw go map. Concurrent
calls to LookupMetadata are therefore unsafe, as concurrent reads from
and writes to the same map are forbidden. This results in runtime panics
in practice.

sync.Map is optimized for maps that only ever grow, and is safe to
access concurrently. Switching the measurementRegistry to a sync.Map
provides for atomic accesses.