DataSubscriber: store pointers to buffers in pool
Slice headers are three words; storing []T in a pool results in
allocations on pool.Get(), since the slice will escape into the
callback.

By storing a pointer to the slice header, and passing that around, we
only need to allocate the slice header on construction of a new slice.