ticks: simplify, add Unix ns conversion funcs, clean comments
Go Time does not support leap seconds, so having the APIs for it doesn't
really accomplish anything. A bunch of the constants are unused, so this
also drops them; redefines PerMillisecond and PerMicrosecond in a more
direct way, but drops the others such as PerDay which are unneeded.

This also adds APIs for converting directly to and from nanosecond
timestamps, to allow bypassing go.Time intermediary conversion.
Converting to and from Unix ns timestamps is just integer math;
converting to and from Go timestamps requires branching and time
zone conversions.

Also removes the comments on the leap flag bits, as whether bits
start at 1 and 0 is contentious and nothing is actually using those
fields except to define the valuemask, the math for which is
straightforward bit manipulation.